requirements.txt
asgiref==3.2.10
certifi==2019.9.11
confusable-homoglyphs==3.2.0
Django==3.0.8
django-address==0.2.4
django-appconf==1.0.4
django-avatar==5.0.0
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-grappelli==2.14.2
django-photologue==3.12
django-profile==0.0.3
django-registration==3.1
django-sortedm2m==3.0.2
ExifRead==2.2.1
Pillow==7.2.0
pipenv==2018.11.26
pytz==2020.1
South==1.0.2
sqlparse==0.3.1
virtualenv==16.7.7
virtualenv-clone==0.5.3
channels==2.4.0
channels-redis==3.0.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
Pillow==7.2.0
psycopg2-binary==2.8.5
pytest-asyncio==0.14.0
pytest-django==3.9.0
